Moved npc related classes to separate package NPCSpawnSpot created
