+++ src/main/java/base/Game.java
@@ -6,6 +6,10 @@ import base.constants.VisibleText;
 import base.events.EventService;
 import base.gameobjects.*;
 import base.gameobjects.interactionzones.InteractionZone;
+import base.gameobjects.npc.Npc;
+import base.gameobjects.npc.NpcLady;
+import base.gameobjects.npc.NpcMan;
+import base.gameobjects.npc.NpcSpot;
 import base.gameobjects.plants.Seed;
 import base.gameobjects.services.*;
 import base.gameobjects.storage.StorageCell;
