+++ src/main/java/base/gameobjects/npc/Npc.java
@@ -1,6 +1,9 @@
-package base.gameobjects;
+package base.gameobjects.npc;
 
 import base.Game;
+import base.gameobjects.AnimatedSprite;
+import base.gameobjects.GameObject;
+import base.gameobjects.Walking;
 import base.gameobjects.interactionzones.InteractionZone;
 import base.graphicsservice.ImageLoader;
 import base.graphicsservice.Rectangle;
